Add files via upload
The security of computer networks has become a crucial issue for businesses, governments, and individuals. Threats to network security are becoming increasingly sophisticated and can have catastrophic consequences for organizations and individuals. Denial of Service (DoS) attacks are one of the most common and damaging threats to computer networks. A DoS attack aims to render a service unavailable to legitimate users by overwhelming the resources of a target system with a massive amount of malicious traffic or requests. These attacks can cause significant financial losses, damage to a company's reputation, and even impact the physical safety of individuals. DoS attacks are constantly evolving, and attackers are using increasingly sophisticated techniques to conceal their identity and make attacks harder to detect.
      The goal of this project is to create a classification model using machine learning techniques that can distinguish between normal and malicious network traffic, while also being able to detect the type of attack based on its network traffic.